prime numbers

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned long long int i, n = 612852475143;
+	unsigned long int i, n = 612852475143;
 
 	for (i = 3; i <= sqrt(n); i += 2)
 	{
@@ -16,6 +16,6 @@ int main(void)
 			n /= i;
 		}
 	}
-	printf("%llu\n", n);
+	printf("%lu\n", n);
 	return (0);
 }
